Grosses modifs
Modification suite à la modif de la web api + ajout de mes annonces

+++ app/src/main/java/com/henallux/projet/smartpark/modele/Parking.java
@@ -24,7 +24,7 @@ public class Parking {
 
     }
 
-    public Parking(int id, String name, String street, String number, String picture, String description, int longitude, int latitude, int idPlace, int idUser, Place place, User user) {
+    public Parking(int id, String name, String street, String number, String picture, String description, double longitude, double latitude, int idPlace, int idUser, Place place, User user) {
         ParkingId = id;
         Name = name;
         Street = street;
